全新改版 1，可以自定义对话框 2，支持kotlin

+++ update-app/src/main/java/com/vector/update_app/DialogActivity.java
@@ -254,14 +254,14 @@ public class DialogActivity extends FragmentActivity implements View.OnClickList
                 }
 
                 @Override
-                public void onProgress(float progress) {
+                public void onProgress(long progress) {
                     if (!DialogActivity.this.isFinishing()) {
                         mNumberProgressBar.setProgress((int) progress);
                     }
                 }
 
                 @Override
-                public void setMax(float total) {
+                public void setMax(long total) {
                     if (!DialogActivity.this.isFinishing()) {
                         mNumberProgressBar.setMax((int) total);
                     }
